[circuit] Rename core to netlist
